requirements.txt
django
boto3
djangorestframework
python-dotenv